fixed h38

+++ src/main/java/com/designpatterns/Hackerrank_8.java
@@ -24,14 +24,14 @@ public class Hackerrank_8 {
 		int[] minInts = Arrays.copyOfRange(arr, 0, 4);
 		int[] maxInts = Arrays.copyOfRange(arr, arr.length - 4, arr.length);
 
-		int minSum = sum(minInts);
-		int maxSum = sum(maxInts);
+		long minSum = sum(minInts);
+		long maxSum = sum(maxInts);
 
 		System.out.println(minSum + " " + maxSum);
 	}
 
-	private static int sum(int[] ints) {
-		int rez = 0;
+	private static long sum(int[] ints) {
+		long rez = 0;
 
 		for (int i = 0; i < ints.length; i++) {
 			rez += ints[i];
